Add files via upload
This is a report on some of our findings on a project where we emphasised on Text Mining and Analysing our findings. We worked on analysis of abstracts of different research journals and how we could use them to predict potential research in that particular field. The report discusses the methods used and the analysis was carried out using R.